The video concern Videable was created, and the valid_video_url? function moved there.
