+++ app/helpers/embed_videos_helper.rb
@@ -31,11 +31,11 @@ module EmbedVideosHelper
     end
   end
 
-  def valid_video_url?
-    return if video_url.blank?
-    return if video_url.match(VIMEO_REGEX)
-    return if video_url.match(YOUTUBE_REGEX)
-    errors.add(:video_url, :invalid)
-  end
+  #  def valid_video_url?
+  #    return if video_url.blank?
+  #    return if video_url.match(VIMEO_REGEX)
+  #    return if video_url.match(YOUTUBE_REGEX)
+  #    errors.add(:video_url, :invalid)
+  #  end
 
 end
